+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# renv (development version)
+
 # renv 1.1.2
 
 * `renv` gains the `sysreqs()` function, which can be used to query the system